refactor!: restrict addon name to valid programming variable name
Previously, any non-empty string could be used as the addon name. To enable the non injection-based
addon registration mechanism in the future, the addon name is now restricted to a valid programming
variable name.

The regular expression is `^[A-Za-z_][A-Za-z0-9_]*$`.

BREAKING CHANGE: The addon name must comply with the regular expression mentioned above.