refactor!: refine extension startup flow (#52)
With the introduction of the on_configure() initialization phase, the startup process of the
extension has been better planned.

- on_configure ~ on_configure_done + on_init ~ on_init_done: Handles its own initialization; cannot
  send or receive messages. The reason for this is that, before `on_init_done`, the extension may
  not be ready to handle external requests, so the received messages need to be temporarily stored.

- ~ on_start: The messages received before on_start() will be temporarily stored, and only after
  on_start() is called will they be sent to the extension. The reason for this is developers
  generally expect `on_start` to occur before any `on_cmd` events.

- on_start ~ on_stop_done: Normal sending and receiving of all messages and results.

BREAKING CHANGE: The initialization actions that were originally performed in on_start() need to be
moved to on_init(). This is because once on_start() is called, on_cmd() will start being received.