feat!: add more codes for standalone testing (#62)
feat!: add more codes for standalone testing

Since the addon for standalone testing is likely not placed within the app's directory tree, the base_dir of the addon is now explicitly specified by the addon itself. Currently, the bindings for C++/Go/Python are designed so that users don't need to explicitly specify the base_dir of the addon. Instead, the addon's register function will attempt to automatically detect the base_dir and pass it to ten_runtime. However, in certain special cases, such as with Cython, the Python code for automatically detecting the addon's base_dir might be optimized away. Therefore, an optional parameter is provided to allow users to explicitly specify the base_dir of the addon.

BREAKING CHANGE: Generally, no changes are required to the addon code, but C++ addons will need to be recompiled.