refactor!: distinguish between graph name and graph id (#68)
refactor!: distinguish between graph name and graph id

Distinguish between the graph name and the graph ID. Each engine started by a graph has a unique graph ID, which is a UUID4. If the graph is a singleton predefined graph, the engine can be found using the singleton predefined graph name.

BREAKING CHANGE: If you want to find a unique engine using the graph name, you need to add "singleton": true in the definition of the predefined graph.